Refactor: Move ShowStmt visitors to AstVisitor
