+++ fe/fe-parser/src/main/java/com/starrocks/sql/ast/ShowEventsStmt.java
@@ -25,6 +25,6 @@ public class ShowEventsStmt extends ShowStmt {
 
     @Override
     public <R, C> R accept(AstVisitor<R, C> visitor, C context) {
-        return ((AstVisitorExtendInterface<R, C>) visitor).visitShowEventStatement(this, context);
+        return visitor.visitShowEventStatement(this, context);
     }
 }
