+++ fe/fe-parser/src/main/java/com/starrocks/sql/ast/ShowPluginsStmt.java
@@ -29,6 +29,6 @@ public class ShowPluginsStmt extends ShowStmt {
 
     @Override
     public <R, C> R accept(AstVisitor<R, C> visitor, C context) {
-        return ((AstVisitorExtendInterface<R, C>) visitor).visitShowPluginsStatement(this, context);
+        return visitor.visitShowPluginsStatement(this, context);
     }
 }
